+++ lexicmap/cmd/search.go
@@ -181,7 +181,7 @@ Output format:
 		// ---------------------------------------------------------------
 
 		if outputLog {
-			log.Infof("LexicMap v%s", VERSION)
+			log.Infof("LexicMap v%s (%s)", VERSION, COMMIT)
 			log.Info("  https://github.com/shenwei356/LexicMap")
 			log.Info()
 		}